[feature] Keep Main class humble by refactoring and moving composables in presentation directory
